More file class refactoring

+++ src/Directory.php
@@ -2,11 +2,45 @@
 
 namespace Nonetallt\File;
 
-use Nonetallt\File\Contract\FilesystemObjectInterface;
+use Nonetallt\File\FilesystemObject;
 use Nonetallt\File\Exception\PermissionException;
+use Nonetallt\File\Exception\FileNotFoundException;
+use Nonetallt\File\Exception\TargetNotDirectoryException;
+use Nonetallt\File\Exception\FilesystemException;
 
-class Directory implements FilesystemObjectInterface
+class Directory extends FilesystemObject
 {
+    /**
+     * Create directory on the filesystem
+     *
+     */
+    public function create(bool $recursive = false)
+    {
+        // Do not attempt creation if file already exists
+        if(file_exists($this->pathname)) {
+            return;
+        }
+
+        // Check if parent directory exists, if not, create it recursively
+        $parentDir = dirname($this->pathname);
+
+        if(! file_exists($parentDir)) {
+            if(! $recursive) {
+                $msg = "Can't create missing parent directory, recursive option not enabled";
+                throw new FilesystemException($msg, $parentDir);
+            }
+
+            (new Directory($parentDir))->create(true);
+        }
+
+        if(! is_writable($parentDir)) {
+            $msg = "Can't create directory, no write permission";
+            throw new PermissionException($msg, $parentDir);
+        }
+
+        mkdir($this->pathname);
+    }
+
     /**
      * Get the size of the objec in filesystem in bytes
      *
@@ -15,8 +49,13 @@ class Directory implements FilesystemObjectInterface
      */
     public function getSize() : int
     {
-        // TODO ?
-        filesize($this->pathname);
+        $size = 0;
+
+        foreach($this->getChildren() as $child)  {
+            $size += $child->getSize();
+        }
+
+        return $size;
     }
 
     /**
@@ -25,7 +64,14 @@ class Directory implements FilesystemObjectInterface
      */
     public function copy(string $destination)
     {
-        copy($this->pathname, $destination);
+        $this->create($destination);
+        var_dump($destination);
+
+        foreach($this->getChildren() as $object) {
+            if($object->isFile() || ($object->isDirectory() && $recursive)) {
+                $object->delete();
+            }
+        }
     }
 
     /**
@@ -35,7 +81,7 @@ class Directory implements FilesystemObjectInterface
     public function move(string $destination)
     {
         $this->copy($destination);
-        $this->delete();
+        $this->delete(true);
     }
 
     /**
@@ -53,8 +99,13 @@ class Directory implements FilesystemObjectInterface
      * Delete filesystem object
      *
      */
-    public function delete()
+    public function delete(bool $recursive = false)
     {
+        if(! $this->isEmpty() && ! $recursive) {
+            $msg = "Directory is not empty, please use the recursive parameter if you wish to delete the directory along with it's contents";
+            throw new FilesystemException($msg, $this->pathname);
+        }
+
         $this->deleteContents(true);
         rmdir($this->pathname);
     }
@@ -67,13 +118,45 @@ class Directory implements FilesystemObjectInterface
      */
     public function deleteContents(bool $recursive = false)
     {
-        $objects = array_diff(scandir($this->pathname), ['.', '..']);
-
-        foreach($objects as $object) {
+        foreach($this->getChildren() as $object) {
             if($object->isFile() || ($object->isDirectory() && $recursive)) {
-                $object->delete();
+                $object->delete(true);
             }
         }
+    }
+
+    /**
+     * Check if the directory is empty
+     *
+     */
+    public function isEmpty() : bool
+    {
+        if(! $this->exists()) {
+            throw new FileNotFoundException($this->pathname);
+        }
+
+        if(! $this->isDirectory()) {
+            throw new TargetNotDirectoryException($this->pathname);
+        }
+
+        return (new \FilesystemIterator($this->pathname))->valid();
+    }
+
+    /**
+     * Get children filesystem objects
+     *
+     */
+    public function getChildren() : array
+    {
+        $children = [];
+
+        foreach(array_diff(scandir($this->pathname), ['.', '..']) as $relativePath) {
+            $realPath = $this->pathname . DIRECTORY_SEPARATOR . $relativePath;
+            $child = is_file($realPath) ? new File($realPath) : new Directory($realPath);
+            $children[] = $child;
+        }
+
+        return $children;
     }
 }
